Move code from main.rs to translator.rs
